Extract htmlToAmp function to files
