+++ src/constants.ts
@@ -1,0 +1,5 @@
+export const height = { name: 'height', value: '1280' }
+export const width = { name: 'width', value: '1920' }
+export const responsive = { name: 'layout', value: 'responsive' }
+export const DEFAULT_HEIGHT = '1280'
+export const DEFAULT_WIDTH = '1920'
