+++ src/index.ts
@@ -1,2 +1,19 @@
 import htmlToAmp from './htmlToAmp'
-export { htmlToAmp }
+import transformFacebook from './transforms/facebook'
+import transformFont from './transforms/fonts'
+import transformIframe from './transforms/iframe'
+import transformImg from './transforms/image'
+import transformInstagram from './transforms/instagram'
+import transformTwitter from './transforms/twitter'
+import transformYoutube from './transforms/youtube'
+
+export {
+	htmlToAmp,
+	transformFacebook,
+	transformFont,
+	transformIframe,
+	transformImg,
+	transformInstagram,
+	transformTwitter,
+	transformYoutube,
+}
