+++ src/transforms/twitter.ts
@@ -1,0 +1,33 @@
+import parse5 from 'parse5'
+import { walk } from '../walk'
+import { getAttribute } from '../domUtils'
+import { responsive, width, height } from '../constants'
+
+export default function transformTwitter(node: parse5.DefaultTreeElement) {
+	const regex = /https:\/\/twitter\.com\/.+\/status[es]*\/([0-9]+).*/
+	return walk(node, async subNode => {
+		const treeNode = subNode as parse5.DefaultTreeElement
+		switch (treeNode.nodeName) {
+			case 'a':
+				if (
+					getAttribute(treeNode, 'href')!.search('https://twitter.com/') === 0
+				) {
+					const tweetUrl = getAttribute(treeNode, 'href') as string
+					const matches = regex.exec(tweetUrl)
+					if (!matches) {
+						return
+					}
+					node.nodeName = 'amp-twitter'
+					node.tagName = 'amp-twitter'
+					node.attrs = [
+						responsive,
+						width,
+						height,
+						{ name: 'data-tweetid', value: matches[1] },
+					]
+					node.childNodes.splice(0, node.childNodes.length)
+				}
+				break
+		}
+	})
+}
